parse-access-log: refactoring and better progress
Yeah, there's a lot of refactoring happening here, but the main thing
is that the SQLite stuff is a lot more reliable, we have well-defined
match items (NamedTuple subclasses), and nice progress reporting with
tqdm if that's avaiable.

Also, data output include IP addresses now, so we can use those for
detecting malicious or misconfigured (or weird) hosts.