countme-update-rawdb.sh + cleanups
This removes daily-run.sh in favor of countme-update-rawdb.sh, which
does the thing that daily-run.sh was _supposed_ to do.

With this commit, all we need to do now is run countme-update-rawdb.sh
and countme-totals.py daily and we'll have nice fresh data whenever we
want it!

At some point we'll probably want to figure out how/when to start
trimming the rawdb but I think we've got at least a year before disk
space becomes an issue.