countme-update-rawdb.sh: handle --start, --all
This commit adds `--start` and `--all` to the `getopt` invocation - they
were listed in the help text and handled in the code, but I forgot to
add them here so the CLI would actually accept them. Fixed.

Now you can manually do partial rawdb runs if needed! Like, say, if you
lost the rawdb during a colo move, but you still have the totalsdb from
before the move, and you just want to parse the logs since the move...