Workaround for "NOT NULL constraint failed" crash (#1)
This commit just makes parse-access-log.py filter out any match item
that has missing (i.e. NULL) field values. This is quite rare, but it
happened once in the Fedora mirror logs from September 16, resulting in
this crash.

Technically, these are still valid `countme` requests, so a future
update should probably remove that constraint, and then re-parse the
logs for September 16 (and any other logs with `countme` items that are
missing `repo=XXX` or `arch=XXX`). But as of now that's 2 hits out of
14.5 million in the logs, so we're not missing much by ignoring
malformed requests like these.