Improve specifying time spans to trim data
This lets users specify that data up to N weeks in the past should be
trimmed (in fact, this is the default mode, the default time span is 13
weeks). Trimming the oldest week is still available, but clamped by the
set duration to avoid trimming too recent data.

Additionally, it doesn’t write to the database by default, this needs
the `--read-write` option added when running the script.

Co-authored-by: James Antill <james@and.org>
Signed-off-by: Nils Philippsen <nils@redhat.com>
Signed-off-by: James Antill <james@and.org>